requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
alembic==1.13.1
alpha-vantage==2.3.1
altair==5.3.0
async-timeout==4.0.3
attrs==23.2.0
blinker==1.8.1
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorlog==6.8.2
contourpy==1.2.1
cycler==0.12.1
filelock==3.14.0
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2024.3.1
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.0.3
idna==3.7
Jinja2==3.1.3
joblib==1.4.2
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
lightgbm==4.3.0
Mako==1.3.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.5
networkx==3.3
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
optuna==3.6.1
packaging==24.0
pandas==2.2.2
pillow==10.3.0
protobuf==4.25.3
pyarrow==16.0.0
pydeck==0.9.0
Pygments==2.17.2
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
requests==2.31.0
rich==13.7.1
rpds-py==0.18.0
scikit-learn==1.4.2
scipy==1.13.0
six==1.16.0
smmap==5.0.1
SQLAlchemy==2.0.29
streamlit==1.34.0
sympy==1.12
tenacity==8.2.3
threadpoolctl==3.5.0
toml==0.10.2
toolz==0.12.1
torch==2.3.0
tornado==6.4
tqdm==4.66.4
triton==2.3.0
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
watchdog==4.0.0
xgboost==2.0.3
yarl==1.9.4
